[2.0] Rebuild worker chroot rpm db only when necessary (#4025)
* rebuild worker chroot rpm db only when necessary

* fail on error. check for any difference in rpm db. remove sudo

* fix typo

+++ toolkit/tools/pkggen/worker/create_worker_chroot.sh
@@ -2,6 +2,9 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
+set -e
+set -o pipefail
+
 # $1 path to create worker base
 # $2 path to text file containing the worker core packages
 # $3 path to find RPMs. May be in PATH/<arch>/*.rpm
@@ -54,32 +57,47 @@ while read -r package || [ -n "$package" ]; do
     install_one_toolchain_rpm "$package"
 done < "$packages"
 
-TEMP_DB_PATH=/temp_db
-echo "Setting up a clean RPM database before the Berkeley DB -> SQLite conversion under '$TEMP_DB_PATH'." | tee -a "$chroot_log"
-chroot "$chroot_builder_folder" mkdir -p "$TEMP_DB_PATH"
-chroot "$chroot_builder_folder" rpm --initdb --dbpath="$TEMP_DB_PATH"
+# These nodes are required in the chroot for certain tools (most importantly, gpg key import when installing 'mariner-repos-shared' package)
+# This is also required to check the rpm db version to see if rebuilding the db is necessary
+mkdir -pv $chroot_builder_folder/dev
+mknod -m 600 $chroot_builder_folder/dev/console c 5 1
+mknod -m 666 $chroot_builder_folder/dev/null c 1 3
+mknod -m 444 $chroot_builder_folder/dev/urandom c 1 9
 
-# Popularing the SQLite database with package info.
-while read -r package || [ -n "$package" ]; do
-    full_rpm_path=$(find "$rpm_path" -name "$package" -type f 2>>"$chroot_log")
-    cp $full_rpm_path $chroot_builder_folder/$package
+# If the host machine rpm version is >= 4.16 (such as Mariner 2.0), it will create an "sqlite" rpm database backend incompatible with Mariner 1.0 (which uses "bdb")
+# To resolve this, enter the 1.0 chroot after the packages are installed, and use the older rpm tool in the chroot to re-create the database in "bdb" format.
+HOST_RPM_VERSION="$(rpm --version)"
+HOST_RPM_DB_BACKEND="$(rpm -E '%{_db_backend}')"
+GUEST_RPM_VERSION="$(chroot "$chroot_builder_folder" rpm --version)"
+GUEST_RPM_DB_BACKEND="$(chroot "$chroot_builder_folder" rpm -E '%{_db_backend}')"
+echo "Current host '$HOST_RPM_VERSION' with rpm db '$HOST_RPM_DB_BACKEND', guest has '$GUEST_RPM_VERSION' with rpm db '$GUEST_RPM_DB_BACKEND'" | tee -a "$chroot_log"
 
-    echo "Adding RPM DB entry to worker chroot: $package." | tee -a "$chroot_log"
+if [[ "$HOST_RPM_DB_BACKEND" == "$GUEST_RPM_DB_BACKEND" ]]; then
+    echo "The host rpm db '$HOST_RPM_DB_BACKEND' matches the guest. Not rebuilding the database." | tee -a "$chroot_log"
+else
+    echo "The host rpm db ('$HOST_RPM_DB_BACKEND') differs from the guest ('$GUEST_RPM_DB_BACKEND'). Rebuilding database for compatibility" | tee -a "$chroot_log"
+    TEMP_DB_PATH="/temp_db"
+    chroot "$chroot_builder_folder" mkdir -p "$TEMP_DB_PATH"
+    chroot "$chroot_builder_folder" rpm --initdb --dbpath="$TEMP_DB_PATH"
+    # Populating the SQLite database with package info.
+    while read -r package || [ -n "$package" ]; do
+        full_rpm_path=$(find "$rpm_path" -name "$package" -type f 2>>"$chroot_log")
+        cp $full_rpm_path $chroot_builder_folder/$package
+        echo "Adding RPM DB entry to worker chroot: $package." | tee -a "$chroot_log"
+        chroot "$chroot_builder_folder" rpm -i -v --nodeps --noorder --force --dbpath="$TEMP_DB_PATH" --justdb "$package" &>> "$chroot_log"
+        chroot "$chroot_builder_folder" rm $package
+    done < "$packages"
+    echo "Overwriting old RPM database with the results of the conversion." | tee -a "$chroot_log"
+    chroot "$chroot_builder_folder" rm -rf /var/lib/rpm
+    chroot "$chroot_builder_folder" mv "$TEMP_DB_PATH" /var/lib/rpm
 
-    chroot "$chroot_builder_folder" rpm -i -v --nodeps --noorder --force --dbpath="$TEMP_DB_PATH" --justdb "$package" &>> "$chroot_log"
-    chroot "$chroot_builder_folder" rm $package
-done < "$packages"
-
-echo "Overwriting old RPM database with the results of the conversion." | tee -a "$chroot_log"
-chroot "$chroot_builder_folder" rm -rf /var/lib/rpm
-chroot "$chroot_builder_folder" mv "$TEMP_DB_PATH" /var/lib/rpm
-
-echo "Importing CBL-Mariner GPG keys." | tee -a "$chroot_log"
-for gpg_key in $(chroot "$chroot_builder_folder" rpm -q -l mariner-repos-shared | grep "rpm-gpg")
-do
-    echo "Importing GPG key: $gpg_key" | tee -a "$chroot_log"
-    chroot "$chroot_builder_folder" rpm --import "$gpg_key"
-done
+    echo "Importing CBL-Mariner GPG keys." | tee -a "$chroot_log"
+    for gpg_key in $(chroot "$chroot_builder_folder" rpm -q -l mariner-repos-shared | grep "rpm-gpg")
+    do
+        echo "Importing GPG key: $gpg_key" | tee -a "$chroot_log"
+        chroot "$chroot_builder_folder" rpm --import "$gpg_key"
+    done
+fi
 
 HOME=$ORIGINAL_HOME
